Defaults: change default line height to 1.35
I chose the second option except I didn't repeat it in the field
description.  The VS Code way of using 0 to "automatically compute the
line height from the font size" seems unnecessary as it implies
something more when it's simply a default ratio of 1.35.  I didn't
repeat it in the field description because all fields should be easy to
reset and having all fields repeat their default would be too noisy.  It
would be better to put a reset to default button next to all fields.