Make side mouse buttons able to jump in code
Currently there's `ctrl+-` and `ctrl+shift+-` to jump forward and back, but ideally this should also be possible with mouse. I initially looked into adding mouse buttons as options for key binds but that seemed like a much taller task than expected, so I settled on simply baking it into the program. This might not be something you want, but it does what it has to so do with this what you will