Fix crash when switching plugin tabs too quickly
I have tried to debug the actual cause but I can't find it.
The only place where `readme_layout` is mutated is within the `event`.
So the only thing I could think of is that the `layout` phase is skipped
somehow?

So simply check if the layout needs to be rebuilt. Else skip.

Signed-off-by: Hanif Ariffin <hanif.ariffin.4326@gmail.com>