Implement UI for renaming files
Implement functionality for renaming files and directories in lapce-app, accessible by right clicking the item in the file explorer and selecting "Rename...".

Backend functionality for file renaming in lapce-proxy was already present and is unchanged.

If the rename fails, the name is not updated in the file explorer and no error is reported.

Open editors are not notified of the rename. If the user saves the contents of an already open editor after performing a rename, the file will be saved in its original location.