Stop returning SyntaxEdit on edits
This was needed so that we can avoid depending on tree-sitter when extracting the buffer + editing logic out to floem-editor
This actually makes it cheaper in the cases where we don't need the edits.
However we are factoring the delta twice now. It would be better to not do that.